requirements.txt
google-genai
rich
python-dotenv
requests
pillow
